requirements.txt
absl-py==2.1.0
albumentations==1.3.1
anykeystore==0.2
apex==0.1
backcall==0.2.0
blessed==1.19.1
boto3==1.17.50
botocore==1.20.50
cachetools==5.4.0
certifi @ file:///croot/certifi_1665076670883/work/certifi
chardet==3.0.4
charset-normalizer==2.1.1
cryptacular==1.6.2
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
einops==0.5.0
et-xmlfile==1.1.0
filelock==3.8.0
fonttools==4.38.0
fsspec==2023.1.0
ftfy==5.1
furl==2.1.3
future==0.18.2
fvcore==0.1.5.post20210413
gensim==4.2.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.32.0
google-auth-oauthlib==0.4.6
googletrans==3.0.0
gpustat==1.0.0
greenlet==3.0.2
grpcio==1.62.2
gym==0.24.1
gym-notices==0.0.7
h11==0.9.0
h2==3.2.0
h5py==3.8.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.16.4
hupper==1.12
hyperframe==5.2.0
idna==2.10
imageio==2.22.2
importlib-metadata==5.0.0
iopath==0.1.7
ipdb==0.13.9
ipython==7.34.0
jedi==0.18.1
jmespath==0.10.0
joblib==1.3.2
kiwisolver @ file:///opt/conda/conda-bld/kiwisolver_1653292039266/work
loguru==0.7.2
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.5.3
matplotlib-inline==0.1.6
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
munch==4.0.0
munkres==1.1.4
networkx==2.6.3
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-ml-py==11.495.46
oauthlib==3.1.0
olefile==0.46
-e git+https://github.com/marlbenchmark/on-policy.git@5778eceff937bfaf7b46cfa6ee20c79ad05b58f5#egg=onpolicy
openai==0.8.0
opencv-python==4.6.0.66
opencv-python-headless==4.8.1.78
openpyxl==3.1.2
orderedmultidict==1.0.1
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.1.5
parso==0.8.3
PasteDeploy==3.1.0
pathtools==0.1.2
pbkdf2==1.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
plaster==1.1.2
plaster-pastedeploy==1.0.1
portalocker==2.3.0
promise==2.3
prompt-toolkit==3.0.31
protobuf==3.20.3
psutil==5.9.2
ptyprocess==0.7.0
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.13.0
PyJWT==2.8.0
pyodbc==4.0.39
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyramid==2.0.2
pyramid-mailer==0.15.1
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python3-openid==3.2.0
pytorch-msssim==0.2.1
pytz==2023.3.post1
PyWavelets==1.3.0
PyYAML==6.0
qudida==0.0.4
regex==2023.12.25
repoze.sendmail==4.4.1
requests==2.28.1
requests-oauthlib==1.3.0
retrying==1.3.4
rfc3986==1.5.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.3.6
safetensors==0.4.1
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
setproctitle==1.2.3
shapely==2.0.4
six @ file:///tmp/build/80754af9/six_1644875935023/work
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.23
tabulate==0.8.9
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
textaugment==2.0.0
textblob==0.17.1
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.6.11
tokenizers==0.13.3
toml==0.10.2
torch==1.12.1+cu113
torchaudio==0.9.0a0+a85b239
torchvision==0.13.1+cu113
tornado @ file:///opt/conda/conda-bld/tornado_1662061693373/work
tqdm==4.66.4
traitlets==5.5.0
transaction==4.0
transformers==4.30.2
translationstring==1.4
typing_extensions==4.7.1
ultralytics==8.0.145
urllib3==1.26.12
velruse==1.1.1
venusian==3.1.0
wcwidth==0.2.5
WebOb==1.8.7
Werkzeug==2.2.3
WTForms==3.0.1
wtforms-recaptcha==0.3.2
yacs==0.1.8
zipp==3.9.0
zope.deprecation==5.0
zope.interface==6.1
zope.sqlalchemy==3.1
